feat(website): Webcam Input is Fully Functional with PoseNet
