--- website/src/Scripts/posenetProcessPoseNew.ts
@@ -1,181 +1,0 @@
-import { PoseNet } from "@tensorflow-models/posenet";
-import * as posenet from "@tensorflow-models/posenet";
-
-function isAndroid() {
-  return /Android/i.test(navigator.userAgent);
-}
-
-function isiOS() {
-  return /iPhone|iPad|iPod/i.test(navigator.userAgent);
-}
-
-export function isMobile() {
-  return isAndroid() || isiOS();
-}
-
-const videoWidth = 600;
-const videoHeight = 500;
-
-/**
- * Loads a the camera to be used in the demo
- *
- */
-async function setupCamera() {
-  if (!navigator.mediaDevices || !navigator.mediaDevices.getUserMedia) {
-    throw new Error(
-      "Browser API navigator.mediaDevices.getUserMedia not available"
-    );
-  }
-
-  const video = document.getElementById("video") as HTMLVideoElement;
-  video.width = videoWidth;
-  video.height = videoHeight;
-
-  const mobile = isMobile();
-  const stream = await navigator.mediaDevices.getUserMedia({
-    audio: false,
-    video: {
-      facingMode: "user",
-      width: mobile ? undefined : videoWidth,
-      height: mobile ? undefined : videoHeight,
-    },
-  });
-  video.srcObject = stream;
-
-  return new Promise<HTMLVideoElement>((resolve) => {
-    video.onloadedmetadata = () => {
-      resolve(video);
-    };
-  });
-}
-
-async function loadVideo() {
-  const video = await setupCamera();
-  video.play();
-
-  return video;
-}
-
-const defaultQuantBytes: any = 2;
-
-const defaultMobileNetMultiplier: any = isMobile() ? 0.5 : 0.75;
-const defaultMobileNetStride: any = 16;
-const defaultMobileNetInputResolution: any = {
-  width: videoWidth,
-  height: videoHeight,
-};
-
-const guiState = {
-  input: {
-    architecture: "MobileNetV1",
-    outputStride: defaultMobileNetStride,
-    inputResolution: defaultMobileNetInputResolution,
-    multiplier: defaultMobileNetMultiplier,
-    quantBytes: defaultQuantBytes,
-  },
-  multiPoseDetection: {
-    maxPoseDetections: 5,
-    minPoseConfidence: 0.15,
-    minPartConfidence: 0.1,
-    nmsRadius: 30.0,
-  },
-  output: {
-    showVideo: true,
-    showSkeleton: true,
-    showPoints: true,
-    showBoundingBox: false,
-  },
-  net: null,
-};
-
-/**
- * Feeds an image to posenet to estimate poses - this is where the magic
- * happens. This function loops with a requestAnimationFrame method.
- */
-function detectPoseInRealTime(video: HTMLVideoElement, net: PoseNet) {
-  /* const canvas = document.getElementById("output");
-  const ctx = canvas.getContext("2d");
-
-  // since images are being fed from a webcam, we want to feed in the
-  // original image and then just flip the keypoints' x coordinates. If instead
-  // we flip the image, then correcting left-right keypoint pairs requires a
-  // permutation on all the keypoints.
-
-  canvas.width = videoWidth;
-  canvas.height = videoHeight; */
-
-  const flipPoseHorizontal = true;
-
-  async function poseDetectionFrame() {
-    let allPoses = await net.estimatePoses(video, {
-      flipHorizontal: flipPoseHorizontal,
-      decodingMethod: "multi-person",
-      maxDetections: guiState.multiPoseDetection.maxPoseDetections,
-      scoreThreshold: guiState.multiPoseDetection.minPartConfidence,
-      nmsRadius: guiState.multiPoseDetection.nmsRadius,
-    });
-
-    console.log(allPoses);
-
-    // Used to draw results; will likely be handled by React in a callback
-    /* let poses = [].concat(allPoses);
-    let minPoseConfidence = +guiState.multiPoseDetection.minPoseConfidence;
-    let minPartConfidence = +guiState.multiPoseDetection.minPartConfidence;
-
-    
-    ctx.clearRect(0, 0, videoWidth, videoHeight);
-
-    if (guiState.output.showVideo) {
-      ctx.save();
-      ctx.scale(-1, 1);
-      ctx.translate(-videoWidth, 0);
-      ctx.drawImage(video, 0, 0, videoWidth, videoHeight);
-      ctx.restore();
-    }
-
-    // For each pose (i.e. person) detected in an image, loop through the poses
-    // and draw the resulting skeleton and keypoints if over certain confidence
-    // scores
-    poses.forEach(({ score, keypoints }) => {
-      if (score >= minPoseConfidence) {
-        if (guiState.output.showPoints) {
-          drawKeypoints(keypoints, minPartConfidence, ctx);
-        }
-        if (guiState.output.showSkeleton) {
-          drawSkeleton(keypoints, minPartConfidence, ctx);
-        }
-        if (guiState.output.showBoundingBox) {
-          drawBoundingBox(keypoints, ctx);
-        }
-      }
-    });*/
-
-    requestAnimationFrame(poseDetectionFrame);
-  }
-
-  poseDetectionFrame();
-}
-
-/**
- * Kicks off the demo by loading the posenet model, finding and loading
- * available camera devices, and setting off the detectPoseInRealTime function.
- */
-export async function bindPage() {
-  const net = await posenet.load({
-    architecture: guiState.input.architecture as any,
-    outputStride: guiState.input.outputStride,
-    inputResolution: guiState.input.inputResolution,
-    multiplier: guiState.input.multiplier,
-    quantBytes: guiState.input.quantBytes,
-  });
-
-  let video;
-
-  try {
-    video = await loadVideo();
-    console.log(video);
-    detectPoseInRealTime(video, net);
-  } catch (e) {
-    console.log(e);
-  }
-}
